requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
python-multipart
pysnmp
pillow
pypdf
PyJWT
passlib